Clean up API and support client-side timeouts.
exec_request now always returns the request object (its first argument). On synchronous calls, request.error will be an exception instance if one occured, else it will be None and result will contain the string result (if any) of the remote API call.  It is None if error is not None.